replace text at render time
