+++ src/components/HTMLEngineProvider/HTMLRenderers/VictoryChartRenderer/components/VictoryChartLabel.tsx
@@ -1,10 +1,14 @@
 import {Skia, Text as SkText} from '@shopify/react-native-skia';
 import type {Color, SkFont} from '@shopify/react-native-skia';
-import React from 'react';
+import React, {useMemo} from 'react';
 import {useChartTypefaces} from '@components/Charts/context/ChartFontsContext';
 import getChartSkiaTypeface from '@components/Charts/utils/getChartSkiaTypeface';
 import type {LabelItem} from '@components/HTMLEngineProvider/HTMLRenderers/VictoryChartRenderer/types';
 import computeTextAnchorPosition from '@components/HTMLEngineProvider/HTMLRenderers/VictoryChartRenderer/utils/computeTextAnchorPosition';
+import {getLocalizedAsOfVictoryChartLabelText} from '@components/HTMLEngineProvider/HTMLRenderers/VictoryChartRenderer/utils/formatAsOfVictoryChartLabel';
+import useCurrentUserPersonalDetails from '@hooks/useCurrentUserPersonalDetails';
+import DateUtils from '@libs/DateUtils';
+import CONST from '@src/CONST';
 
 type VictoryChartLabelsProps = LabelItem;
 
@@ -23,7 +27,13 @@ type ProcessedLine = {
  */
 function VictoryChartLabel({x, y, text, color, fontSize, fontWeight, fontFamily, fontStyle, lineHeight, textAnchor = 'start', verticalAnchor = 'middle'}: VictoryChartLabelsProps) {
     const typefaces = useChartTypefaces();
-    const processedLines = text.split('\n').reduce(
+    const currentUserPersonalDetails = useCurrentUserPersonalDetails();
+    const viewerTimezone = useMemo(() => {
+        const timezone = currentUserPersonalDetails?.timezone ?? CONST.DEFAULT_TIME_ZONE;
+        return DateUtils.getCurrentTimezone(timezone).selected;
+    }, [currentUserPersonalDetails?.timezone?.automatic, currentUserPersonalDetails?.timezone?.selected]);
+    const displayText = useMemo(() => getLocalizedAsOfVictoryChartLabelText(text, viewerTimezone), [text, viewerTimezone]);
+    const processedLines = displayText.split('\n').reduce(
         (acc, line, index) => {
             const lineColor = color?.[index];
             const lineFontSize = fontSize?.[index];
