+++ src/components/HTMLEngineProvider/HTMLRenderers/VictoryChartRenderer/parsers/processVictoryChartTree.ts
@@ -3,14 +3,13 @@ import lodashMerge from 'lodash/merge';
 import type {TNode} from 'react-native-render-html';
 import {X_KEY} from '@components/HTMLEngineProvider/HTMLRenderers/VictoryChartRenderer/constants';
 import type {ProcessNodeResult} from '@components/HTMLEngineProvider/HTMLRenderers/VictoryChartRenderer/types';
-import type {SelectedTimezone} from '@src/types/onyx/PersonalDetails';
 import PARSER_REGISTRY from './parserRegistry';
 
 /**
  * Recursively walk the HTML tnode tree, dispatching each node to its registered parser
  * and merging the results into a single chart config.
  */
-function processVictoryChartTree(tnode: TNode, typeface: SkTypeface | null, rootProcessedResult: ProcessNodeResult | null, viewerTimezone?: SelectedTimezone): ProcessNodeResult {
+function processVictoryChartTree(tnode: TNode, typeface: SkTypeface | null, rootProcessedResult: ProcessNodeResult | null): ProcessNodeResult {
     const data: ProcessNodeResult['data'] = {};
     const yKeys: ProcessNodeResult['yKeys'] = [];
     let xAxis: ProcessNodeResult['xAxis'];
@@ -25,7 +24,7 @@ function processVictoryChartTree(tnode: TNode, typeface: SkTypeface | null, root
 
     const parser = PARSER_REGISTRY[tnode.tagName ?? ''];
     if (parser) {
-        const result = parser(tnode, typeface, rootProcessedResult, viewerTimezone);
+        const result = parser(tnode, typeface, rootProcessedResult);
         if (result.data) {
             lodashMerge(data, result.data);
         }
@@ -65,7 +64,7 @@ function processVictoryChartTree(tnode: TNode, typeface: SkTypeface | null, root
     const rootProcessedNodeResult = rootProcessedResult ?? {data, xKey: X_KEY, yKeys, xAxis, yAxis, domain, domainPadding, padding, isHorizontal, categories, labelItems, legendItems};
 
     for (const child of tnode.children) {
-        const childResult = processVictoryChartTree(child, typeface, rootProcessedNodeResult, viewerTimezone);
+        const childResult = processVictoryChartTree(child, typeface, rootProcessedNodeResult);
         lodashMerge(data, childResult.data);
         yKeys.push(...childResult.yKeys);
         if (childResult.xAxis) {
